io: resurrected sendfile() support and some more cleanups along the way
* socket_sink now properly accelerates file source transmission via sendfile()-syscall
* sink::pump() returns ssize_t instead of size_t to reflect -1 (errno) properly
* file_source: improved API, added access to passed file_ptr
* file_source: removed redundant ::close(fd_) as ~file() already ::close()'s the fd

+++ src/x0/io/async_writer.hpp
@@ -59,7 +59,7 @@ public:
 private:
 	void write()
 	{
-		int rv = sink_->pump(*source_);
+		ssize_t rv = sink_->pump(*source_);
 
 		if (rv == 0)
 		{
